keyboard: Emit key release right away
At the moment the on-screen keyboard doesn't emit the key release event
until the user stops pushing the keyboard button with their pointer.

This means if the user uses the pointer to hold the button down, it can
generate repeat events for some keys.

But this creates a bit of an inconsistency in behavior between keys that
support multiple choices via long press and those that don't. The ones
that support long press, don't repeat, instead they show the available
choices.

Furthermore, key repeat doesn't work for any of the keys with the
wayland backend, since key repeat is a client side thing, and we just
don't have it implemented for this path.

Also, key repeat is repeating the wrong keys right now, even on X11, for
keys that require a shift level (see
https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/2045 )

Given key repeat is a dubious feature in an on-screen keyboard to begin
with, and it's only implemented for one backend, and it's not even
completely working on that backend, it's probably best to drop support.

This commit changes the on-screen keyboard to always emit a key
release immediately after each key press.